+++ gcovr/formats/base.py
@@ -39,6 +39,7 @@ class BaseHandler:
             "sort_uncovered",
             "sort_percent",
             "sort_reverse",
+            "sort_casefold",
             "search_paths",
             "source_encoding",
             "starting_dir",