+++ gcovr/formats/html/write.py
@@ -429,6 +429,7 @@ def write_report(covdata: CovData, output_file: str, options: Options) -> None:
         by_num_uncovered=options.sort_uncovered,
         by_percent_uncovered=options.sort_percent,
         reverse=options.sort_reverse,
+        casefold=options.sort_casefold,
         filename_uses_relative_pathname=True,
     )
 
@@ -669,6 +670,7 @@ def write_directory_pages(
             by_num_uncovered=options.sort_uncovered,
             by_percent_uncovered=options.sort_percent,
             reverse=options.sort_reverse,
+            casefold=options.sort_casefold,
             filename_uses_relative_pathname=True,
         )