+++ gcovr/formats/json/write.py
@@ -102,6 +102,7 @@ def write_summary_report(covdata, output_file: str, options: Options):
         by_num_uncovered=options.sort_uncovered,
         by_percent_uncovered=options.sort_percent,
         reverse=options.sort_reverse,
+        casefold=options.sort_casefold,
     )
 
     for key in keys: